refactor(ui): split large editor components
