+++ Facio/Models/Timesheet.swift
@@ -1,208 +1,5 @@
 import Foundation
 
-// MARK: - Jour de la semaine
-
-enum JourSemaine: Int, Codable, CaseIterable, Identifiable {
-    case lundi = 0, mardi, mercredi, jeudi, vendredi, samedi, dimanche
-
-    var id: Int { rawValue }
-
-    var label: String {
-        switch self {
-        case .lundi: return "Lundi"
-        case .mardi: return "Mardi"
-        case .mercredi: return "Mercredi"
-        case .jeudi: return "Jeudi"
-        case .vendredi: return "Vendredi"
-        case .samedi: return "Samedi"
-        case .dimanche: return "Dimanche"
-        }
-    }
-
-    var shortLabel: String {
-        switch self {
-        case .lundi: return "Lun"
-        case .mardi: return "Mar"
-        case .mercredi: return "Mer"
-        case .jeudi: return "Jeu"
-        case .vendredi: return "Ven"
-        case .samedi: return "Sam"
-        case .dimanche: return "Dim"
-        }
-    }
-
-    func label(for lang: AppLanguage) -> String {
-        L10n.weekdayLabel(rawValue, lang)
-    }
-
-    func shortLabel(for lang: AppLanguage) -> String {
-        L10n.weekdayShort(rawValue, lang)
-    }
-}
-
-// MARK: - Entree journaliere
-
-struct TimesheetDay: Identifiable, Codable, Hashable {
-    var id: UUID = UUID()
-    /// Date du jour (sans heure)
-    var dateString: String  // "2026-03-01" — stocke comme string pour serialisation fiable
-    var heures: Decimal = 0
-
-    var date: Date {
-        TimesheetDay.dateFormatter.date(from: dateString) ?? Date()
-    }
-
-    var jourSemaine: JourSemaine {
-        let cal = Calendar(identifier: .gregorian)
-        var weekday = cal.component(.weekday, from: date) - 2  // dimanche=1 en Calendar, on veut lundi=0
-        if weekday < 0 { weekday = 6 }
-        return JourSemaine(rawValue: weekday) ?? .lundi
-    }
-
-    /// Jour du mois (ex: "26", "1", "15")
-    var jourDuMois: Int {
-        Calendar(identifier: .gregorian).component(.day, from: date)
-    }
-
-    /// Mois (1-12)
-    var mois: Int {
-        Calendar(identifier: .gregorian).component(.month, from: date)
-    }
-
-    static let dateFormatter: DateFormatter = {
-        let f = DateFormatter()
-        f.dateFormat = "yyyy-MM-dd"
-        f.locale = Locale(identifier: "fr_FR")
-        return f
-    }()
-
-    init(date: Date, heures: Decimal = 0) {
-        self.id = UUID()
-        self.dateString = TimesheetDay.dateFormatter.string(from: date)
-        self.heures = heures
-    }
-
-    // MARK: - Codable (backwards-compatible)
-
-    enum CodingKeys: String, CodingKey {
-        case id, dateString, date, heures
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        id = container.decodeOrDefault(UUID.self, forKey: .id, default: UUID())
-        if let persistedDateString = try container.decodeIfPresent(String.self, forKey: .dateString),
-           !persistedDateString.isEmpty {
-            dateString = persistedDateString
-        } else if let persistedDate = try container.decodeIfPresent(Date.self, forKey: .date) {
-            dateString = TimesheetDay.dateFormatter.string(from: persistedDate)
-        } else {
-            dateString = TimesheetDay.dateFormatter.string(from: Date())
-        }
-        heures = container.decodeOrDefault(Decimal.self, forKey: .heures, default: 0)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encode(id, forKey: .id)
-        try container.encode(dateString, forKey: .dateString)
-        try container.encode(heures, forKey: .heures)
-    }
-}
-
-// MARK: - Semaine calendaire
-
-struct TimesheetWeek: Identifiable, Codable, Hashable {
-    var id: UUID = UUID()
-    var numero: Int = 1
-    /// 7 jours (lundi a dimanche) avec leurs vraies dates
-    var jours: [TimesheetDay] = []
-
-    init() {}
-
-    /// Total heures de la semaine
-    var totalHeures: Decimal {
-        jours.reduce(0) { $0 + $1.heures }
-    }
-
-    func heuresNormales(seuil: Decimal = 35) -> Decimal {
-        min(totalHeures, seuil)
-    }
-
-    func heuresSupplementaires(seuil: Decimal = 35) -> Decimal {
-        max(totalHeures - seuil, 0)
-    }
-
-    /// Calcule les heures sup attribuables au mois courant dans cette semaine,
-    /// en tenant compte des heures des jours d'un mois adjacent (chronologiquement).
-    /// Les jours sont ordonnés lun→dim. Les premières `seuil` heures sont normales,
-    /// le reste est overtime, réparti par mois selon l'ordre chronologique.
-    func heuresSupPourMois(moisPeriode: Int, seuil: Decimal = 35, adjacentHours: [String: Decimal]) -> Decimal {
-        var heuresPrecedentes: Decimal = 0
-        var heuresMoisCourant: Decimal = 0
-        var moisCourantCommence = false
-
-        for jour in jours {
-            let h: Decimal
-            if jour.mois == moisPeriode {
-                h = jour.heures
-                heuresMoisCourant += h
-                moisCourantCommence = true
-            } else {
-                h = adjacentHours[jour.dateString] ?? jour.heures
-                if !moisCourantCommence {
-                    // Jours d'un autre mois avant le mois courant
-                    heuresPrecedentes += h
-                }
-                // Jours après le mois courant n'affectent pas l'overtime de CE mois
-            }
-        }
-        return max(0, heuresPrecedentes + heuresMoisCourant - seuil) - max(0, heuresPrecedentes - seuil)
-    }
-
-    /// Heures normales du mois courant dans cette semaine (cross-période)
-    func heuresNormalesPourMois(moisPeriode: Int, seuil: Decimal = 35, adjacentHours: [String: Decimal]) -> Decimal {
-        let heuresMois = jours.filter { $0.mois == moisPeriode }.reduce(Decimal(0)) { $0 + $1.heures }
-        let sup = heuresSupPourMois(moisPeriode: moisPeriode, seuil: seuil, adjacentHours: adjacentHours)
-        return heuresMois - sup
-    }
-
-    /// Le lundi de cette semaine
-    var dateDebut: Date? {
-        jours.first?.date
-    }
-
-    /// Label : "26 fev - 02 mar"
-    var label: String {
-        guard let first = jours.first, let last = jours.last else { return "Semaine \(numero)" }
-        let f = DateFormatter()
-        f.dateFormat = "dd MMM"
-        f.locale = Locale(identifier: "fr_FR")
-        return "\(f.string(from: first.date)) — \(f.string(from: last.date))"
-    }
-
-    func label(for lang: AppLanguage) -> String {
-        guard let first = jours.first, let last = jours.last else { return L10n.week(lang, number: numero) }
-        let f = DateFormatter()
-        f.dateFormat = "dd MMM"
-        f.locale = Locale(identifier: lang == .fr ? "fr_FR" : "en_US")
-        return "\(f.string(from: first.date)) — \(f.string(from: last.date))"
-    }
-
-    // MARK: - Codable
-
-    enum CodingKeys: String, CodingKey {
-        case id, numero, jours
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        id = container.decodeOrDefault(UUID.self, forKey: .id, default: UUID())
-        numero = container.decodeOrDefault(Int.self, forKey: .numero, default: 1)
-        jours = container.decodeOrDefault([TimesheetDay].self, forKey: .jours, default: [])
-    }
-}
-
 // MARK: - Periode (mois)
 
 @Observable
