+++ Facio/Services/SyncService.swift
@@ -1,174 +1,5 @@
 import Foundation
 import Observation
-import Security
-
-// MARK: - Sync State
-
-struct SyncState: Codable {
-    var documentsDirty: Bool = false
-    var clientsDirty: Bool = false
-    var companyDirty: Bool = false
-    var timesheetsDirty: Bool = false
-    var lastFullSyncAt: Date?
-    var migrationCompleted: Bool = false
-
-    var hasDirtyData: Bool {
-        documentsDirty || clientsDirty || companyDirty || timesheetsDirty
-    }
-
-    enum CodingKeys: String, CodingKey {
-        case documentsDirty, clientsDirty, companyDirty, timesheetsDirty, lastFullSyncAt, migrationCompleted
-    }
-
-    init(
-        documentsDirty: Bool = false,
-        clientsDirty: Bool = false,
-        companyDirty: Bool = false,
-        timesheetsDirty: Bool = false,
-        lastFullSyncAt: Date? = nil,
-        migrationCompleted: Bool = false
-    ) {
-        self.documentsDirty = documentsDirty
-        self.clientsDirty = clientsDirty
-        self.companyDirty = companyDirty
-        self.timesheetsDirty = timesheetsDirty
-        self.lastFullSyncAt = lastFullSyncAt
-        self.migrationCompleted = migrationCompleted
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        documentsDirty = container.decodeOrDefault(Bool.self, forKey: .documentsDirty, default: false)
-        clientsDirty = container.decodeOrDefault(Bool.self, forKey: .clientsDirty, default: false)
-        companyDirty = container.decodeOrDefault(Bool.self, forKey: .companyDirty, default: false)
-        timesheetsDirty = container.decodeOrDefault(Bool.self, forKey: .timesheetsDirty, default: false)
-        lastFullSyncAt = try container.decodeIfPresent(Date.self, forKey: .lastFullSyncAt)
-        migrationCompleted = container.decodeOrDefault(Bool.self, forKey: .migrationCompleted, default: false)
-    }
-}
-
-// MARK: - Sync Config
-
-struct SyncConfig {
-    // DB partagee par defaut — lue depuis ~/.facio_config
-    static var defaultURL: String { Secrets.supabaseURL }
-    static var defaultAPIKey: String { Secrets.supabaseAnonKey }
-
-    // Cles UserDefaults
-    private static let customURLKey = "supabase_custom_url"
-    private static let customAPIKeyKey = "supabase_custom_api_key"
-    private static let useCustomKey = "supabase_use_custom"
-    private static let enabledKey = "sync_enabled"
-
-    /// Sync activee
-    static var isEnabled: Bool {
-        get { UserDefaults.standard.bool(forKey: enabledKey) }
-        set { UserDefaults.standard.set(newValue, forKey: enabledKey) }
-    }
-
-    /// Utiliser une DB custom au lieu de la DB partagee
-    static var useCustomDB: Bool {
-        get { UserDefaults.standard.bool(forKey: useCustomKey) }
-        set { UserDefaults.standard.set(newValue, forKey: useCustomKey) }
-    }
-
-    /// URL custom
-    static var customURL: String {
-        get { UserDefaults.standard.string(forKey: customURLKey) ?? "" }
-        set { UserDefaults.standard.set(newValue, forKey: customURLKey) }
-    }
-
-    /// API key custom
-    static var customAPIKey: String {
-        get {
-            if let keychainValue = SyncKeychain.string(for: customAPIKeyKey) {
-                return keychainValue
-            }
-            guard let legacyValue = UserDefaults.standard.string(forKey: customAPIKeyKey),
-                  !legacyValue.isEmpty
-            else { return "" }
-            guard SyncKeychain.set(legacyValue, for: customAPIKeyKey) else { return "" }
-            UserDefaults.standard.removeObject(forKey: customAPIKeyKey)
-            return legacyValue
-        }
-        set {
-            if newValue.isEmpty {
-                SyncKeychain.delete(customAPIKeyKey)
-                UserDefaults.standard.removeObject(forKey: customAPIKeyKey)
-            } else if SyncKeychain.set(newValue, for: customAPIKeyKey) {
-                UserDefaults.standard.removeObject(forKey: customAPIKeyKey)
-            } else {
-                UserDefaults.standard.removeObject(forKey: customAPIKeyKey)
-            }
-        }
-    }
-
-    /// URL effective (custom ou par defaut)
-    static var url: String {
-        useCustomDB && !customURL.isEmpty ? customURL : defaultURL
-    }
-
-    /// API key effective
-    static var apiKey: String {
-        useCustomDB && !customAPIKey.isEmpty ? customAPIKey : defaultAPIKey
-    }
-
-    static var isConfigured: Bool {
-        !url.isEmpty && !apiKey.isEmpty
-    }
-}
-
-private enum SyncKeychain {
-    private static let service = "app.facio.sync"
-
-    static func string(for account: String) -> String? {
-        var query = baseQuery(account: account)
-        query[kSecReturnData as String] = true
-        query[kSecMatchLimit as String] = kSecMatchLimitOne
-
-        var item: CFTypeRef?
-        let status = SecItemCopyMatching(query as CFDictionary, &item)
-        guard status == errSecSuccess,
-              let data = item as? Data
-        else { return nil }
-        return String(data: data, encoding: .utf8)
-    }
-
-    @discardableResult
-    static func set(_ value: String, for account: String) -> Bool {
-        guard let data = value.data(using: .utf8) else { return false }
-        let query = baseQuery(account: account)
-        let update = [kSecValueData as String: data]
-        let status = SecItemUpdate(query as CFDictionary, update as CFDictionary)
-        if status == errSecSuccess { return true }
-        guard status == errSecItemNotFound else { return false }
-
-        var add = query
-        add[kSecValueData as String] = data
-        add[kSecAttrAccessible as String] = kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly
-        return SecItemAdd(add as CFDictionary, nil) == errSecSuccess
-    }
-
-    static func delete(_ account: String) {
-        SecItemDelete(baseQuery(account: account) as CFDictionary)
-    }
-
-    private static func baseQuery(account: String) -> [String: Any] {
-        [
-            kSecClass as String: kSecClassGenericPassword,
-            kSecAttrService as String: service,
-            kSecAttrAccount as String: account,
-        ]
-    }
-}
-
-extension TimesheetWeek {
-    init(id: UUID = UUID(), numero: Int = 1, jours: [TimesheetDay] = []) {
-        self.id = id
-        self.numero = numero
-        self.jours = jours
-    }
-}
 
 // MARK: - Sync Service
